signalement des commentaires
signalement des commentaires mis en place 'flagComment'

+++ src/controller/BackController.php
@@ -4,44 +4,66 @@ namespace App\src\controller;
 
 use App\config\Parameter;
 
+
 class BackController extends Controller
 {
+    private function checkAdmin()
+    {
+        if($this->session->get('pseudo')!='admin'){
+            header('Location:../public/index.php?route=connexionUser');
+        }
+
+    }
+
     public function addArticle(Parameter $post)
     {
-        if($post->get('submit')) {
-            $this->articleDAO->addArticle($post);
-            $this->session->set('add_article', 'Le nouvel article a bien été ajouté');
-            header('Location: ../public/index.php');
+        if(!$this->checkAdmin()){
+            if($post->get('submit')) {
+                $this->articleDAO->addArticle($post);
+                $this->session->set('add_article', 'Le nouvel article a bien été ajouté');
+                header('Location: ../public/index.php');
+            }
+            return $this->view->render('add_article', [
+                'post' => $post
+            ]);
         }
-        return $this->view->render('add_article', [
-            'post' => $post
-        ]);
+
     }
 
     public function deleteArticle($articleId)
     {
-        $this->commentDAO->deleteComments($articleId);
-        $this->articleDAO->deleteArticle($articleId);
-        $this->session->set('delete_article', 'L\'article a bien été supprimé');
-        header('Location: ../public/index.php');
+        if(!$this->checkAdmin()){
+
+            $this->commentDAO->deleteComments($articleId);
+            $this->articleDAO->deleteArticle($articleId);
+            $this->session->set('delete_article', 'L\'article a bien été supprimé');
+            header('Location: ../public/index.php');
+        }
+
+
     }
 
 
     public function modifyArticle(Parameter $post, $id)
     {
-        $article = $this->articleDAO->getArticle($id);
+        if(!$this->checkAdmin()){
+            $article = $this->articleDAO->getArticle($id);
 
-        //var_dump($post);
+            //var_dump($post);
 
-        if($post->get('submit')) {
-            $this->articleDAO->updateArticle($post, $id);
-            $this->session->set('modify_article', 'L\'article a bien été modifié');
-            header('Location: ../public/index.php?route=article&articleId='.$id);
+            if($post->get('submit')) {
+                $this->articleDAO->updateArticle($post, $id);
+                $this->session->set('modify_article', 'L\'article a bien été modifié');
+                header('Location: ../public/index.php?route=article&articleId='.$id);
+            }
+
+            return $this->view->render('modify', [
+                'article' => $article
+            ]);
+
         }
 
-        return $this->view->render('modify', [
-            'article' => $article
-        ]);
+
 
     }
 
@@ -62,29 +84,41 @@ class BackController extends Controller
         ]);
     }
 
+
+
     public function suppComment($id)
     {
-        $article_id = $this->commentDAO->getArticleIdForComment($id);
-        $this->commentDAO->deleteComment($id);
-        header('Location: ../public/index.php?route=article&articleId='.$article_id[0]);
+        if(!$this->checkAdmin()){
+
+                $article_id = $this->commentDAO->getArticleIdForComment($id);
+                $this->commentDAO->deleteComment($id);
+                header('Location: ../public/index.php?route=article&articleId='.$article_id[0]);
+
+
+        }
+
+
     }
 
     public function modifyComment(Parameter $post, $id)
     {
-        //récupérer le commentaire
-        $comment = $this->commentDAO->getComment($id);
+        if(!$this->checkAdmin()){
+            //récupérer le commentaire
+            $comment = $this->commentDAO->getComment($id);
 
-        if($post->get('submit')) {
-            $this->commentDAO->modifyComment($post, $id);
-            $article_id = $this->commentDAO->getArticleIdForComment($id);
-            header('Location: ../public/index.php?route=article&articleId='.$article_id[0]);
+            if($post->get('submit')) {
+                $this->commentDAO->modifyComment($post, $id);
+                $article_id = $this->commentDAO->getArticleIdForComment($id);
+                header('Location: ../public/index.php?route=article&articleId='.$article_id[0]);
+            }
+
+            return $this->view->render('modifyComment', [
+                'comment' => $comment
+            ]);
+
         }
 
-        return $this->view->render('modifyComment', [
-            'comment' => $comment
-        ]);
     }
-
 
     //pour les utilisateurs
